+++ data/src/main/java/com/example/data/movies/source/local/model/RemoteKeys.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class RemoteKeys(
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo(name = "movie_id")
-    val movieID: Int,
+    val movieID: Long,
     val prevKey: Int?,
     val currentPage: Int,
     val nextKey: Int?,
