+++ domain/src/main/java/com/example/domain/moviedetails/usecase/GetMovieDetailsUseCase.kt
@@ -3,8 +3,10 @@ package com.example.domain.moviedetails.usecase
 import com.example.domain.moviedetails.repository.MovieDetailsRepository
 import javax.inject.Inject
 
-class GetMovieDetailsUseCase @Inject constructor( private val movieDetailsRepository: MovieDetailsRepository
+class GetMovieDetailsUseCase @Inject constructor(
+    private val movieDetailsRepository: MovieDetailsRepository
 ) {
 
-    suspend fun loadMovieDetails(movieId: Int) = movieDetailsRepository.loadMovieDetails(movieId)
+    suspend fun loadMovieDetails(movieId: Long) = movieDetailsRepository
+        .loadMovieDetails(movieId)
 }
