+++ presentation/src/main/java/com/example/presentation/movies/contract/MoviesContract.kt
@@ -25,7 +25,7 @@ class MoviesContract {
 
     sealed class Effect : ViewSideEffect {
         sealed class Navigation : Effect() {
-            data class ToMovieDetails(val movieId: Int) : Effect()
+            data class ToMovieDetails(val movieId: Long) : Effect()
         }
     }
